chore: fixed warrior bot behavior; added weapon info on the screen

+++ client/internal/entity/player_stats.go
@@ -4,17 +4,17 @@ type PlayerStats interface {
 	Class() Class
 	Radius() float32
 	Speed() float32
-	MaxHP() int64
+	MaxHP() int32
 }
 
 type playerStats struct {
 	class  Class
 	radius float32
 	speed  float32
-	maxHP  int64
+	maxHP  int32
 }
 
-func NewPlayerStats(class Class, radius, speed float64, maxHP int64) PlayerStats {
+func NewPlayerStats(class Class, radius, speed float64, maxHP int32) PlayerStats {
 	return &playerStats{
 		class:  class,
 		radius: float32(radius),
@@ -35,6 +35,6 @@ func (s *playerStats) Speed() float32 {
 	return s.speed
 }
 
-func (s *playerStats) MaxHP() int64 {
+func (s *playerStats) MaxHP() int32 {
 	return s.maxHP
 }
